% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-json-control.R
\name{import_control_json}
\alias{import_control_json}
\title{Import a control JSON file and standardise its contents
ready for use as read function arguments}
\usage{
import_control_json(file)
}
\arguments{
\item{file}{string location of the file}
}
\value{
list object
}
\description{
Import a control JSON file and standardise its contents
ready for use as read function arguments
}
