% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise-colnames.R
\name{sanitise_colnames}
\alias{sanitise_colnames}
\title{Sanitise colname vector to lower camel case}
\usage{
sanitise_colnames(name_vec)
}
\arguments{
\item{name_vec}{character vector of column names}
}
\value{
character vector in lower camel case
}
\description{
Sanitise colname vector to lower camel case
}
