% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise-colnames.R
\name{sanitise_df_colnames}
\alias{sanitise_df_colnames}
\title{Sanitse all column names in a dataframe to lower camel case}
\usage{
sanitise_df_colnames(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with renamed columns
}
\description{
Sanitse all column names in a dataframe to lower camel case
}
