% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-json-control.R
\name{transform_json_for_control}
\alias{transform_json_for_control}
\title{Transform json list item to argument-ready data structure}
\usage{
transform_json_for_control(json_list_item, item_name = "unnamed")
}
\arguments{
\item{json_list_item}{Item ready from jsonlite}

\item{item_name}{Name of item}
}
\value{
value
}
\description{
Transform json list item to argument-ready data structure
}
